+++ CHANGELOG.markdown
@@ -6,6 +6,7 @@
 
 - Support for progress bar
 - Support for push/pop item width
+- Support for ID stack manipulation (integer values)
 - `ImVec4::zero()`
 - `Into` array and tuple conversions for ImVec2 and ImVec4